+++ ArcSlider/ArcSlider.js
@@ -211,7 +211,12 @@ const ArcSliderBase = kind({
 
 	render: ({'aria-valuetext': ariaValuetext, backgroundColor, componentRef, disabled, endAngle, foregroundColor, isFocused, max, min, radius, size, slotCenter, startAngle, strokeWidth, value, ...rest}, context) => {
 		const {accent: accentColor} = context || {};
-		const valueAngle = valueToAngle(value, min, max, startAngle, endAngle);
+		let valueAngle;
+		if(max <= min) {
+			valueAngle = valueToAngle(min, min, min, startAngle, endAngle);
+		} else {
+		  valueAngle = valueToAngle(value, min, max, startAngle, endAngle);
+		}
 		const knobPosition = angleToPosition(valueAngle, radius - (strokeWidth / 2), size);
 
 		delete rest.step;