+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The following is a curated list of changes in the Enact agate module, newest cha
 - `agate/LabeledIconButton` `max-width` to display huge sized icon correctly
 - `agate/MediaPlayer` previous button functionality to play media from the beginning after being paused
 - `agate/TimePicker` transition direction for meridiem when hour is changed
+- `agate/ArcSlider` to display correctly when `min` prop equals `max` prop or when `min` is greater than `max`
 
 ## [1.1.1] - 2020-12-23